Mod: search_controller
