+++ app/views/places/result.html.erb
