+++ app/views/places/search.html.erb
