+++ config/routes.rb
@@ -12,8 +12,8 @@ Rails.application.routes.draw do
   # get 'search', to: 'searches#search', as: 'search'
   # post 'search', to: 'searches#search'
   # get 'result', to: 'searches#result', as: 'result'
-  get 'index', to: 'searches#index'
-  get 'search', to: 'searches#search', as: 'search_places'
+  get 'index', to: 'places#index'
+  get 'search', to: 'places#search', as: 'search_places'
   # Define your application routes per the DSL in https://guides.rubyonrails.org/routing.html
 
   # Defines the root path route ("/")
